% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countiteration.R
\name{counting_iteration}
\alias{counting_iteration}
\title{Counting iteration}
\usage{
counting_iteration(data, x)
}
\arguments{
\item{data}{is a dataframe or a tibble}

\item{x}{is the name of the variable for which we want to count the number of iteration for the different possibilities}
}
\value{
a plot of columns in the decreasing order, showing the counts for every possibilities of the variable of interest
}
\description{
This function count the number of iteration for the different possibilities of one variable of a table
}
\examples{
counting_iteration(palmerpenguins::penguins, species)
counting_iteration(dplyr::starwars, homeworld)

}
